PlayList

+++ app/src/main/java/com/legend/ace18/songs/utils/SongUtils.java
@@ -9,7 +9,7 @@ public class SongUtils {
      * Timer Format
      * Hours:Minutes:Seconds
      * */
-    public String milliSecondsToTimer(long milliseconds){
+    public String milliSecondsToTimer(int milliseconds){
         String finalTimerString = "";
         String secondsString = "";
